#105 Simplified Shared components into a single main view with one MainViewModel
