--- src/AdminAssistant.Blazor/Client/Shared/Breadcrumb/BreadcrumbComponent.razor
@@ -1,9 +1,0 @@
-@inherits AdminAssistantComponentBase<AdminAssistant.UI.Shared.Breadcrumb.IBreadcrumbViewModel>
-
-<div class="cl-breadcrumb">
-  <ol class="breadcrumb">
-    <li class="breadcrumb-item active"><i class="fa fa-home" aria-hidden="true"></i>&nbsp;Home</li>
-    <li class="breadcrumb-item active"><i class="fa @this.vm.ActiveMode.Icon" aria-hidden="true"></i>&nbsp;@this.vm.ActiveMode.Label</li>
-    <li class="breadcrumb-item active"><i class="fa @this.vm.ActiveModule.Icon" aria-hidden="true"></i>&nbsp;@this.vm.ActiveModule.Label</li>
-  </ol>
-</div>
