--- src/AdminAssistant.Blazor/Client/Shared/Footer/FooterComponent.razor
@@ -1,5 +1,0 @@
-@inherits AdminAssistantComponentBase<AdminAssistant.UI.Shared.Footer.IFooterViewModel>
-
-<div class="cl-footer">
-  <div>@this.vm.FooterText</div>
-</div>
