--- src/AdminAssistant.Blazor/Client/Shared/Header/HeaderComponent.razor
@@ -1,23 +1,0 @@
-@inherits AdminAssistantComponentBase<AdminAssistant.UI.Shared.Header.IHeaderViewModel>
-
-<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
-
-  <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbar2" aria-controls="navbar2" aria-expanded="false" aria-label="Toggle navigation">
-    <span class="navbar-toggler-icon"></span>
-  </button>
-  <a class="navbar-brand" href="#">Admin Assistant</a>
-  <div class="collapse navbar-collapse justify-content-between" id="navbar2">
-
-    <form class="form-inline my-2 my-lg-0">
-      <input class="form-control mr-sm-2" type="text" placeholder="Search">
-      <button class="btn btn-secondary my-2 my-sm-0" type="submit">Search</button>
-    </form>
-
-    <div>
-      <h4><i aria-hidden="true" class="fa @this.vm.ActiveMode.Icon"></i>&nbsp;@this.vm.ActiveMode.Label&nbsp;-&nbsp;<i aria-hidden="true" class="fa @this.vm.ActiveModule.Icon"></i>&nbsp;@this.vm.ActiveModule.Label</h4>
-    </div>
-
-    <div>
-    </div>
-  </div>
-</nav>
