+++ src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/MainWindow.xaml
@@ -2,35 +2,68 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:sf="http://schemas.syncfusion.com/wpf"
+        xmlns:sfnav="http://schemas.syncfusion.com/wpf"
         xmlns:sfskin ="clr-namespace:Syncfusion.SfSkinManager;assembly=Syncfusion.SfSkinManager.WPF"
-        xmlns:header="clr-namespace:AdminAssistant.WPF.Shared.Header"
-        xmlns:sidebar="clr-namespace:AdminAssistant.WPF.Shared.Sidebar"
-        xmlns:breadcrumb="clr-namespace:AdminAssistant.WPF.Shared.Breadcrumb"
-        xmlns:footer="clr-namespace:AdminAssistant.WPF.Shared.Footer"
         xmlns:dashboardModule="clr-namespace:AdminAssistant.WPF.Modules.DashboardModule"
         xmlns:accountsModule="clr-namespace:AdminAssistant.WPF.Modules.AccountsModule"
-        xmlns:shared="clr-namespace:AdminAssistant.UI.Shared;assembly=AdminAssistant"
         sfskin:SfSkinManager.VisualStyle="Office2019Black"
         Title="Admin Assistant" Height="720" Width="1280"
         DataContext="{Binding MainWindowViewModel, Source={StaticResource ViewModelLocator}}">
 
   <Window.Resources>
-
     <DataTemplate x:Key="DashboardViewTemplate" DataType="{x:Type UserControl}">
       <dashboardModule:DashboardComponent />
     </DataTemplate>
-
     <DataTemplate x:Key="AccountsViewTemplate" DataType="{x:Type UserControl}">
       <accountsModule:AccountsComponent />
     </DataTemplate>
   </Window.Resources>
 
   <DockPanel LastChildFill="True">
-    <header:HeaderComponent DockPanel.Dock="Top" />
-    <footer:FooterComponent DockPanel.Dock="Bottom" />
-    <sidebar:SidebarComponent DockPanel.Dock="Left" />
-    <breadcrumb:BreadcrumbComponent DockPanel.Dock="Top" />
+
+    <!-- Start Header -->
+    <Grid DockPanel.Dock="Top">
+      <Label Content="Header" />
+    </Grid>
+    <!-- End Header -->
     
+    <!-- Start Footer -->
+    <Grid DockPanel.Dock="Bottom">
+      <Label Height="auto" Content="{Binding FooterText}" HorizontalAlignment="Center" />
+    </Grid>
+    <!-- End Footer -->
+
+    <!-- Start Sidebar -->
+    <Grid DockPanel.Dock="Left">
+      <Grid.RowDefinitions>
+        <RowDefinition x:Name="navBar" Height="auto" />
+        <RowDefinition x:Name="overflow" Height="*" />
+      </Grid.RowDefinitions>
+      <sf:SfTreeNavigator Grid.Row="0" Width="100" ItemsSource="{Binding Modules}" >
+        <sf:SfTreeNavigator.HeaderTemplate>
+          <DataTemplate>
+            <ComboBox ItemsSource="{Binding Path=Modes}" DisplayMemberPath="Label" SelectedValuePath="Label" SelectedValue="{Binding Path=ActiveMode}" />
+          </DataTemplate>
+        </sf:SfTreeNavigator.HeaderTemplate>
+        <sf:SfTreeNavigator.ItemTemplate>
+          <HierarchicalDataTemplate ItemsSource="{Binding Modules}">
+            <StackPanel Orientation="Horizontal">
+              <TextBlock Text="{Binding Label}" FontWeight="Bold" VerticalAlignment="Center" Margin="18 0 0 0"/>
+            </StackPanel>
+          </HierarchicalDataTemplate>
+        </sf:SfTreeNavigator.ItemTemplate>
+      </sf:SfTreeNavigator>
+    </Grid>
+    <!-- End Sidebar -->
+
+    <!-- Start Breadcrumb -->
+    <Grid DockPanel.Dock="Top">
+      <Label Content="Breadcrumb" />
+    </Grid>
+    <!-- End Breadcrumb -->
+
+
+    <!-- Start Main Content -->
     <ContentControl Content="{Binding}">
       <ContentControl.Style>
         <Style TargetType="{x:Type ContentControl}">
@@ -46,7 +79,7 @@
         </Style>
       </ContentControl.Style>
     </ContentControl>
-    <ContentControl x:Name="ContentPannel" Content="{Binding SelectedViewModel}" />
+    <!--<ContentControl x:Name="ContentPannel" Content="{Binding SelectedViewModel}" />-->
+    <!-- End Main Content -->
   </DockPanel>
-  
 </sf:ChromelessWindow>
