--- src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/Shared/Breadcrumb/BreadcrumbComponent.xaml
@@ -1,8 +1,0 @@
-<UserControl x:Class="AdminAssistant.WPF.Shared.Breadcrumb.BreadcrumbComponent"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             DataContext="{Binding BreadcrumbViewModel, Source={StaticResource ViewModelLocator}}">
-  <Grid>
-    <Label Content="Breadcrumb" />
-  </Grid>
-</UserControl>
