--- src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/Shared/Breadcrumb/BreadcrumbComponent.xaml.cs
@@ -1,19 +1,0 @@
-using System.Windows.Controls;
-using AdminAssistant.UI.Shared.Breadcrumb;
-
-namespace AdminAssistant.WPF.Shared.Breadcrumb
-{
-    public partial class BreadcrumbComponent : UserControl
-    {
-        public BreadcrumbComponent()
-        {
-            this.InitializeComponent();
-        }
-
-        public BreadcrumbComponent(IBreadcrumbViewModel breadcrumbViewModel)
-            : this()
-        {
-            this.DataContext = breadcrumbViewModel;
-        }
-    }
-}
