--- src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/Shared/Footer/FooterComponent.xaml.cs
@@ -1,12 +1,0 @@
-using System.Windows.Controls;
-
-namespace AdminAssistant.WPF.Shared.Footer
-{
-    public partial class FooterComponent : UserControl
-    {
-        public FooterComponent()
-        {
-            InitializeComponent();
-        }
-    }
-}
