--- src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/Shared/Header/HeaderComponent.xaml
@@ -1,8 +1,0 @@
-<UserControl x:Class="AdminAssistant.WPF.Shared.Header.HeaderComponent"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             DataContext="{Binding HeaderViewModel, Source={StaticResource ViewModelLocator}}">
-  <Grid>
-    <Label Content="Header" />
-  </Grid>
-</UserControl>
