--- src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/Shared/Header/HeaderComponent.xaml.cs
@@ -1,12 +1,0 @@
-using System.Windows.Controls;
-
-namespace AdminAssistant.WPF.Shared.Header
-{
-    public partial class HeaderComponent : UserControl
-    {
-        public HeaderComponent()
-        {
-            InitializeComponent();
-        }
-    }
-}
