--- src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/Shared/Sidebar/SidebarComponent.xaml
@@ -1,45 +1,0 @@
-<UserControl x:Class="AdminAssistant.WPF.Shared.Sidebar.SidebarComponent"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:sfnav="http://schemas.syncfusion.com/wpf"
-             DataContext="{Binding SidebarViewModel, Source={StaticResource ViewModelLocator}}">
-  <Grid>
-    <Grid.RowDefinitions>
-      <RowDefinition x:Name="navBar" Height="auto" />
-      <RowDefinition x:Name="overflow" Height="*" />
-    </Grid.RowDefinitions>
-
-    <sfnav:SfTreeNavigator Grid.Row="0" Width="100" SelectionChanged="{Binding }}" ItemsSource="{Binding Modules}" >
-      <sfnav:SfTreeNavigator.HeaderTemplate>
-        <DataTemplate>
-          <ComboBox ItemsSource="{Binding Path=Modes}" DisplayMemberPath="Label" SelectedValuePath="Label" SelectedValue="{Binding Path=ActiveMode}" />
-
-          <!--<ComboBox SelectedItem="Company" SelectedValue="Company">
-            <ComboBoxItem>Company</ComboBoxItem>
-            <ComboBoxItem>Personal</ComboBoxItem>
-          </ComboBox>-->
-        </DataTemplate>
-      </sfnav:SfTreeNavigator.HeaderTemplate>
-      <sfnav:SfTreeNavigator.ItemTemplate>
-        <HierarchicalDataTemplate ItemsSource="{Binding Modules}">
-          <StackPanel Orientation="Horizontal">
-            <TextBlock Text="{Binding Label}" FontWeight="Bold" VerticalAlignment="Center" Margin="18 0 0 0"/>
-          </StackPanel>
-        </HierarchicalDataTemplate>
-      </sfnav:SfTreeNavigator.ItemTemplate>
-      <!--<sfnav:SfTreeNavigatorItem Header="Dashboard"/>
-      <sfnav:SfTreeNavigatorItem Header="Mail"/>
-      <sfnav:SfTreeNavigatorItem Header="Calendar"/>
-      <sfnav:SfTreeNavigatorItem Header="Contacts"/>
-      <sfnav:SfTreeNavigatorItem Header="Tasks"/>
-      <sfnav:SfTreeNavigatorItem Header="Accounts"/>
-      <sfnav:SfTreeNavigatorItem Header="Assets"/>
-      <sfnav:SfTreeNavigatorItem Header="Billing"/>
-      <sfnav:SfTreeNavigatorItem Header="Budget"/>
-      <sfnav:SfTreeNavigatorItem Header="Documents"/>
-      <sfnav:SfTreeNavigatorItem Header="Reports"/>
-      <sfnav:SfTreeNavigatorItem Header="Admin"/>-->
-    </sfnav:SfTreeNavigator>
-    
-  </Grid>
-</UserControl>
