--- src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/Shared/Sidebar/SidebarComponent.xaml.cs
@@ -1,12 +1,0 @@
-using System.Windows.Controls;
-
-namespace AdminAssistant.WPF.Shared.Sidebar
-{
-    public partial class SidebarComponent : UserControl
-    {
-        public SidebarComponent()
-        {
-            this.InitializeComponent();
-        }
-    }
-}
