+++ src/AdminAssistant.UI/Shared/AppService.cs
