+++ src/AdminAssistant.UI/Shared/MainWindowViewModel.cs
@@ -1,6 +1,7 @@
+using System.Collections.Generic;
 using AdminAssistant.Framework.Providers;
-using AdminAssistant.UI.Modules.AccountsModule;
-using AdminAssistant.UI.Modules.DashboardModule;
+using AdminAssistant.UI.Services;
+using AdminAssistant.UI.Shared.Sidebar;
 using Microsoft.Toolkit.Mvvm.Messaging;
 
 namespace AdminAssistant.UI.Shared
@@ -8,17 +9,52 @@ namespace AdminAssistant.UI.Shared
     [System.Diagnostics.CodeAnalysis.SuppressMessage("Build", "CA1812", Justification = "Compiler dosen't understand dependency injection")]
     internal class MainWindowViewModel : ViewModelBase, IMainWindowViewModel
     {
-        private readonly IMessenger messenger;
+        private const string SelectedModuleStyle = "active";
 
-        public MainWindowViewModel(IMessenger messenger, ILoggingProvider loggerProvider)
+        private readonly IMessenger messenger;
+        private readonly IAppService appService;
+
+        private readonly SidebarStateSettings contractedSidebarState;
+        private readonly SidebarStateSettings expandedSidebarState;
+        private readonly ModeSelectionStateSettings contractedModeSelectionDropDownState;
+        private readonly ModeSelectionStateSettings expandedModeSelectionDropDownState;
+
+        public MainWindowViewModel(IMessenger messenger, IAppService appService, ILoggingProvider loggerProvider)
             : base(loggerProvider)
         {
             this.messenger = messenger;
             this.messenger.RegisterAll(this);
 
+            this.appService = appService;
+
             this.SelectedModule = ModuleEnum.Dashboard;
+
+            this.activeMode = appService.GetDefaultMode();
+            this.activeModule = appService.GetDefaultModule();
+
+            this.Log.Start();
+            this.FooterText = $"Admin Assistant - V{System.Reflection.Assembly.GetExecutingAssembly().GetName().Version}";
+            this.Log.Finish();
+
+
+            this.contractedSidebarState = new SidebarStateSettings(ExpandedContractedStateToggle.Contracted, "fa fa-lg fa-angle-double-right", "cl-navbar-contracted", false);
+            this.expandedSidebarState = new SidebarStateSettings(ExpandedContractedStateToggle.Expanded, "fa fa-lg fa-angle-double-left", "cl-navbar-expanded", true);
+            this.Sidebar = this.expandedSidebarState;
+
+            this.contractedModeSelectionDropDownState = new ModeSelectionStateSettings(ExpandedContractedStateToggle.Contracted, string.Empty);
+            this.expandedModeSelectionDropDownState = new ModeSelectionStateSettings(ExpandedContractedStateToggle.Expanded, "show");
+            this.ModeSelectionDropDown = this.contractedModeSelectionDropDownState;
+
+            this.Modes = this.appService.GetModes();
+            this.ActiveMode = this.appService.GetDefaultMode();
+
+            this.Modules = this.appService.GetModules();
+            this.ActiveModule = this.appService.GetDefaultModule();
+            this.ActiveModule.StyleClass = SelectedModuleStyle;
         }
         ~MainWindowViewModel() => this.messenger.UnregisterAll(this);
+
+        public string FooterText { get; }
 
         private ModuleEnum selectedModule;
         public ModuleEnum SelectedModule
@@ -27,6 +63,75 @@ namespace AdminAssistant.UI.Shared
             private set => SetProperty(ref selectedModule, value);
         }
 
-        public void Receive(ModuleSelectionChangedMessage message) => this.SelectedModule = message.Value.Module;
+        private ModeSelectionItem activeMode;
+        public ModeSelectionItem ActiveMode
+        {
+            get => activeMode;
+            set => this.SetProperty(ref activeMode, value);
+        }
+
+        private ModuleSelectionItem activeModule;
+        public ModuleSelectionItem ActiveModule
+        {
+            get => activeModule;
+            set => this.SetProperty(ref activeModule, value);
+        }
+
+        public SidebarStateSettings Sidebar { get; private set; }
+        public ModeSelectionStateSettings ModeSelectionDropDown { get; private set; }
+
+        public List<ModeSelectionItem> Modes { get; private set; }
+
+        public List<ModuleSelectionItem> Modules { get; private set; }
+
+        public void OnModeSelectionDropDownClick()
+        {
+            switch (this.ModeSelectionDropDown.State)
+            {
+                case ExpandedContractedStateToggle.Contracted:
+                    this.ModeSelectionDropDown = this.expandedModeSelectionDropDownState;
+                    break;
+
+                case ExpandedContractedStateToggle.Expanded:
+                    this.ModeSelectionDropDown = this.contractedModeSelectionDropDownState;
+                    break;
+            }
+        }
+
+        public void OnSideBarControlButtonClick()
+        {
+            switch (this.Sidebar.State)
+            {
+                case ExpandedContractedStateToggle.Contracted:
+                    this.Sidebar = this.expandedSidebarState;
+                    this.Modes.ForEach((module) => { module.Label = module.Tag; });
+                    this.Modules.ForEach((module) => { module.Label = module.Tag; });
+                    break;
+
+                case ExpandedContractedStateToggle.Expanded:
+                    this.Sidebar = this.contractedSidebarState;
+                    this.Modes.ForEach((module) => { module.Label = string.Empty; });
+                    this.Modules.ForEach((module) => { module.Label = string.Empty; });
+                    break;
+            }
+            this.OnPropertyChanged(nameof(this.Sidebar));
+        }
+
+        public void OnSelectedModeChanged(ModeSelectionItem selectedMode)
+        {
+            this.ActiveMode = selectedMode;
+            this.ModeSelectionDropDown = this.contractedModeSelectionDropDownState;
+
+            this.messenger.Send(new ModeSelectionChangedMessage(this.ActiveMode));
+        }
+
+        public void OnSelectedModuleChanged(ModuleSelectionItem selectedModule)
+        {
+            this.ActiveModule = selectedModule;
+            this.Modules.ForEach(x => x.StyleClass = string.Empty);
+            this.ActiveModule.StyleClass = SelectedModuleStyle;
+
+            this.messenger.Send(new ModuleSelectionChangedMessage(this.ActiveModule));
+        }
     }
 }
