+++ src/AdminAssistant/AdminAssistant.csproj
@@ -60,7 +60,6 @@
     <Folder Include="DomainModel\Modules\MailModule\" />
     <Folder Include="DomainModel\Modules\TasksModule\" />
     <Folder Include="DomainModel\Modules\ReportsModule\" />
-    <Folder Include="UI\Shared\Services\" />
   </ItemGroup>
 
   <ItemGroup>
