--- src/AdminAssistant/UI/Shared/Breadcrumb/IBreadcrumbViewModel.cs
@@ -1,11 +1,0 @@
-using Microsoft.Toolkit.Mvvm.Messaging;
-
-namespace AdminAssistant.UI.Shared.Breadcrumb
-{
-    public interface IBreadcrumbViewModel : IViewModelBase, IRecipient<ModeSelectionChangedMessage>, IRecipient<ModuleSelectionChangedMessage>
-    {
-        ModeSelectionItem ActiveMode { get; }
-
-        ModuleSelectionItem ActiveModule { get; }
-    }
-}
