--- src/AdminAssistant/UI/Shared/Footer/IFooterViewModel.cs
@@ -1,7 +1,0 @@
-namespace AdminAssistant.UI.Shared.Footer
-{
-    public interface IFooterViewModel : IViewModelBase
-    {
-        string FooterText { get; }
-    }
-}
