--- src/AdminAssistant/UI/Shared/Header/IHeaderViewModel.cs
@@ -1,10 +1,0 @@
-using Microsoft.Toolkit.Mvvm.Messaging;
-
-namespace AdminAssistant.UI.Shared.Header
-{
-    public interface IHeaderViewModel : IViewModelBase, IRecipient<ModeSelectionChangedMessage>, IRecipient<ModuleSelectionChangedMessage>
-    {
-        ModeSelectionItem ActiveMode { get; }
-        ModuleSelectionItem ActiveModule { get; }
-    }
-}
