+++ src/AdminAssistant/UI/Shared/IAppService.cs
