+++ src/AdminAssistant/UI/Shared/IMainWindowViewModel.cs
@@ -1,9 +1,26 @@
-using Microsoft.Toolkit.Mvvm.Messaging;
+using System.Collections.Generic;
+using AdminAssistant.UI.Shared.Sidebar;
 
 namespace AdminAssistant.UI.Shared
 {
-    public interface IMainWindowViewModel : IViewModelBase, IRecipient<ModuleSelectionChangedMessage>
+    public interface IMainWindowViewModel : IViewModelBase
     {
-        ModuleEnum SelectedModule { get; }
+        SidebarStateSettings Sidebar { get; }
+
+        ModeSelectionStateSettings ModeSelectionDropDown { get; }
+
+        List<ModeSelectionItem> Modes { get; }
+        ModeSelectionItem ActiveMode { get; }
+
+        List<ModuleSelectionItem> Modules { get; }
+        ModuleSelectionItem ActiveModule { get; }
+
+        void OnModeSelectionDropDownClick();
+        void OnSideBarControlButtonClick();
+
+        void OnSelectedModeChanged(ModeSelectionItem mode);
+        void OnSelectedModuleChanged(ModuleSelectionItem selectedModule);
+
+        string FooterText { get; }
     }
 }
