+++ src/AdminAssistant/UI/Shared/ModeSelectionSettings.cs
