--- src/AdminAssistant/UI/Shared/Services/IAppStateStore.cs
@@ -1,17 +1,0 @@
-//using System;
-//using AdminAssistant.UI.Shared;
-
-//namespace AdminAssistant.UI.Services
-//{
-//    public interface IAppStateStore
-//    {
-//        event Action? ActiveModeChanged;
-//        event Action? ActiveModuleChanged;
-
-//        ModeSelectionItem ActiveMode { get; }
-//        ModuleSelectionItem ActiveModule { get; }
-
-//        void OnActiveModeChanged(ModeSelectionItem mode);
-//        void OnActiveModuleChanged(ModuleSelectionItem module);
-//    }
-//}
