--- src/AdminAssistant/UI/Shared/Sidebar/ISidebarViewModel.cs
@@ -1,23 +1,0 @@
-using System.Collections.Generic;
-
-namespace AdminAssistant.UI.Shared.Sidebar
-{
-    public interface ISidebarViewModel : IViewModelBase
-    {
-        SidebarStateSettings Sidebar { get; }
-
-        ModeSelectionStateSettings ModeSelectionDropDown { get; }
-
-        List<ModeSelectionItem> Modes { get; }
-        ModeSelectionItem ActiveMode { get; }
-
-        List<ModuleSelectionItem> Modules { get; }
-        ModuleSelectionItem ActiveModule { get; }
-
-        void OnModeSelectionDropDownClick();
-        void OnSideBarControlButtonClick();
-
-        void OnSelectedModeChanged(ModeSelectionItem mode);
-        void OnSelectedModuleChanged(ModuleSelectionItem selectedModule);
-    }
-}
