+++ src/AdminAssistant/UI/Shared/SidebarStateSettings.cs
